Initial commit to add social media icons to template
Add a new region to the page template so a block of HTML containing
the images can be added and configured to only show on the ferries
homepage.

Not sure I am totally happy with this as I would rather the HTML
block was baked directly into the template itself.