Update form alter function to reflect changed form name
During early development I had to change the name of the home page
'Find your sailing schedule by route' function because of a duplicate
function needed on the 'Schedules' landing page. I forgot to update
the function name which alters the select route function to unset the
fields which are passed in the URL by default.